requirements.txt
# Fed-ProFiLA-AD Requirements
# 联邦学习声学异常检测项目依赖

# Core dependencies
torch>=1.12.0
torchaudio>=0.12.0
numpy>=1.21.0
scipy>=1.8.0

# Data processing
pandas>=1.4.0
scikit-learn>=1.1.0

# Configuration
PyYAML>=6.0

# Audio processing
librosa>=0.9.0
soundfile>=0.10.0

# Visualization
matplotlib>=3.5.0
seaborn>=0.11.0
tqdm>=4.64.0

# Logging and utilities
tqdm>=4.64.0
wandb>=0.12.0  # Optional: for experiment tracking

# Development dependencies (optional)
pytest>=7.0.0
black>=22.0.0
flake8>=4.0.0
